+++ src/ui/svelte/ProgressView/ProgressFolder.svelte
@@ -4,12 +4,16 @@
 	import { RECORDING, WSFolder } from "src/model/folder";
 	import { FormatNumber, FormatWords, MoveTheTarget } from "src/util";
 	import { onDestroy, onMount } from "svelte";
+	import CachedStatsDisplay from "../CachedStatsDisplay.svelte";
 	import ProgressBar from "./ProgressBar.svelte";
 
 	export let plugin: WordStatisticsPlugin;
 	export let events = plugin.events;
 
 	export let folder: WSFolder;
+	export let showStats: boolean = true;
+
+	let statsDisplay: CachedStatsDisplay;
 
 	let progress: ProgressBar;
 	let goal: number;
@@ -33,9 +37,15 @@
 		events.on(WSEvents.Folder.Renamed, onFolderUpdate, { filter: null });
 		events.on(WSEvents.Folder.GoalSet, onFolderUpdate, { filter: null });
 		events.on(WSEvents.Folder.TitleSet, onFolderUpdate, { filter: null });
-		events.on(WSEvents.Setting.Recording, onSettingUpdate, { filter: null});
-		events.on(WSEvents.Setting.MovingTarget, onSettingUpdate, { filter: null});
-		events.on(WSEvents.Folder.RecordingSet, onFolderUpdate, { filter: null });
+		events.on(WSEvents.Setting.Recording, onSettingUpdate, {
+			filter: null,
+		});
+		events.on(WSEvents.Setting.MovingTarget, onSettingUpdate, {
+			filter: null,
+		});
+		events.on(WSEvents.Folder.RecordingSet, onFolderUpdate, {
+			filter: null,
+		});
 		updateAll();
 	});
 
@@ -46,9 +56,15 @@
 		events.off(WSEvents.Folder.Renamed, onFolderUpdate, { filter: null });
 		events.off(WSEvents.Folder.GoalSet, onFolderUpdate, { filter: null });
 		events.off(WSEvents.Folder.TitleSet, onFolderUpdate, { filter: null });
-		events.off(WSEvents.Setting.Recording, onSettingUpdate, { filter: null});
-		events.off(WSEvents.Setting.MovingTarget, onSettingUpdate, { filter: null});
-		events.off(WSEvents.Folder.RecordingSet, onFolderUpdate, { filter: null });
+		events.off(WSEvents.Setting.Recording, onSettingUpdate, {
+			filter: null,
+		});
+		events.off(WSEvents.Setting.MovingTarget, onSettingUpdate, {
+			filter: null,
+		});
+		events.off(WSEvents.Folder.RecordingSet, onFolderUpdate, {
+			filter: null,
+		});
 	});
 
 	function onSettingUpdate(event: WSSettingEvent) {
@@ -76,7 +92,10 @@
 			goalText = FormatWords(goal);
 			countText = goal > 0 ? FormatNumber(count) : FormatWords(count);
 			label = goal > 0 ? countText + " / " + goalText : countText;
-			progress?.SetProgress(goal > 0 ? (count / goal) * 100 : 0, goal === 0);
+			progress?.SetProgress(
+				goal > 0 ? (count / goal) * 100 : 0,
+				goal === 0
+			);
 			inherit = folder.recording === RECORDING.INHERIT;
 			recording = folder.isRecording;
 		} else {
@@ -85,13 +104,26 @@
 		}
 		folder = folder;
 	}
+
+	$: if (showStats && folder instanceof WSFolder && folder.duration > 0) {
+		statsDisplay?.Update(folder);
+	}
 </script>
 
 <div class="ws-progress-folder">
-	<h2><span class="record" class:recording={recording && !inherit} class:inherit={recording && inherit}/>{title}</h2>
+	<h2>
+		<span
+			class="record"
+			class:recording={recording && !inherit}
+			class:inherit={recording && inherit}
+		/>{title}
+	</h2>
 	{#if title != name}
 		<h3 class="ws-progress-folder-name">({name})</h3>
 	{/if}
 	<ProgressBar bind:this={progress} />
 	<div class="ws-progress-label">{label}</div>
+	{#if showStats && folder instanceof WSFolder && folder.duration > 0}
+		<CachedStatsDisplay updateObject={folder} bind:this={statsDisplay} />
+	{/if}
 </div>