+++ src/ui/svelte/StatsForDayDisplay.svelte
@@ -2,7 +2,7 @@
 	import { DateTime } from "luxon";
 	import type { WSFileStat } from "src/model/file";
 	import { WordStats } from "src/model/stats";
-	import { FormatWords, SecondsToHMS } from "src/util";
+	import { FormatNumber, FormatWords, MillisecondsToReadableDuration } from "src/util";
 	import { onMount } from "svelte";
 
 	export let stats: WSFileStat[];
@@ -48,11 +48,11 @@
 		wordsExported = WordStats.GetWordsExported(stats);
 		writingTime = WordStats.GetWritingTime(stats);
 
-		WPM = Math.round(netWords / (duration / 60000));
-		WAPM = Math.round(wordsAdded / (duration / 60000));
+		WPM = Math.round(((netWords / (duration / 60000)) + Number.EPSILON) * 100) / 100;
+		WAPM = Math.round(((wordsAdded / (duration / 60000)) + Number.EPSILON) * 100) / 100
 
-		WPMA = Math.round(netWords / (writingTime / 60000));
-		WAPMA = Math.round(wordsAdded / (writingTime / 60000));
+		WPMA = Math.round(((netWords / (writingTime / 60000)) + Number.EPSILON) * 100) / 100;
+		WAPMA = Math.round(((wordsAdded / (writingTime / 60000)) + Number.EPSILON) * 100) / 100
 	}
 
 	$: okay =
@@ -69,10 +69,16 @@
 		<div class="value">
 			{endDate?.toLocaleString(DateTime.DATETIME_SHORT)}
 		</div>
+		<div class="heading">Duration</div>
+		<div class="value">{MillisecondsToReadableDuration(duration)}</div>
+		<div class="heading">Writing Time</div>
+		<div class="value">{MillisecondsToReadableDuration(writingTime)}</div>
 		<div class="heading">Start Words</div>
 		<div class="value">{FormatWords(startWords)}</div>
 		<div class="heading">End Words</div>
 		<div class="value">{FormatWords(endWords)}</div>
+		<div class="heading">Net Words</div>
+		<div class="value">{FormatWords(netWords)}</div>
 		<div class="heading">Words Added</div>
 		<div class="value">{FormatWords(wordsAdded)}</div>
 		<div class="heading">Words Deleted</div>
@@ -81,13 +87,9 @@
 		<div class="value">{FormatWords(wordsImported)}</div>
 		<div class="heading">Words Exported</div>
 		<div class="value">{FormatWords(wordsExported)}</div>
-		<div class="heading">Duration</div>
-		<div class="value">{SecondsToHMS(duration / 1000)}</div>
-		<div class="heading">Writing Time</div>
-		<div class="value">{SecondsToHMS(writingTime / 1000)}</div>
 		<div class="heading">WPM</div>
-		<div class="value">{FormatWords(WPM)}</div>
+		<div class="value">{`${FormatNumber(WPM)} wpm`}</div>
 		<div class="heading">WAPM</div>
-		<div class="value">{FormatWords(WAPM)}</div>
+		<div class="value">{`${FormatNumber(WAPM)} wapm`}</div>
 	</div>
 {/if}